Move ListComponent in its own class
